// src/modules/users/entities/enums/enums.ts
export enum Gender {
  MALE = 'Male',
  FEMALE = 'Female',
  PREFER_NOT_TO_SAY = 'Prefer not to say',
}

export enum ExperienceLevel {
  BEGINNER = 'Beginner/Novice',
  INTERMEDIATE = 'Intermediate',
  UPPER_INTERMEDIATE = 'Upper Intermediate',
  EXPERIENCED = 'Skillful and Experienced',
}

export enum SkillLevel {
  BEGINNER = 'Beginner',
  INTERMEDIATE = 'intermediate',
  ADVANCED = 'Advanced',
}

export enum PlayerPosition {
  GOALKEEPER = 'Goalkeeper',
  DEFENDER = 'Defender',
  MIDFIELDER = 'Midfielder',
  STRIKER = 'Striker',
}

export enum AvailableDays {
  MONDAY = 'Monday',
  TUESDAY = 'Tuesday',
  WEDNESDAY = 'Wednesday',
  THURSDAY = 'Thursday',
  FRIDAY = 'Friday',
  SATURDAY = 'Saturday',
  SUNDAY = 'Sunday',
}

export enum PreferredTime {
  MORNING = 'Morning',
  AFTERNOON = 'Afternoon',
  EVENING = 'Evening',
}

export enum PaymentStatus {
  PENDING = 'Pending',
  PAID = 'Paid',
  FAILED = 'Failed',
}

export enum SubscriptionPlan {
  free = 'free',
  U5_U8 = 'U5_U8',
  U9_U12 = 'U9_U12',
  U13_U14 = 'U13_U14',
  U15_U18 = 'U15_U18',
}

export enum TShirtSize {
  YXXS = 'YXXS',
  YXS = 'YXS',
  YS = 'YS',
  YM = 'YM',
  YL = 'YL',
  YXL = 'YXL',
  XS = 'XS',
  S = 'S',
  M = 'M',
  L = 'L',
  XL = 'XL',
  XXL = 'XXL',
}